requirements.txt
joblib
matplotlib
numpy
pandas
Pillow
scikit_learn
streamlit
streamlit_option_menu
tensorflow
tensorflow_intel
